+++ src/MotoHealth.Core/Bot/AccidentReporting/AccidentReporter.cs
@@ -2,13 +2,13 @@
 {
     public sealed class AccidentReporter
     {
-        public AccidentReporter(int telegramUserId, string phoneNumber)
+        public AccidentReporter(long telegramUserId, string phoneNumber)
         {
             TelegramUserId = telegramUserId;
             PhoneNumber = phoneNumber;
         }
 
-        public int TelegramUserId { get; }
+        public long TelegramUserId { get; }
 
         public string PhoneNumber { get; }
     }
